this is hard

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	long long int d;
 	float e;
 	{
-	printf("Size of a char: %ld byte(s)\n", (sizeof(a)));
-	printf("Size of an int: %li byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %li bytes(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %ld bytes(s)\n", (unsigned long) sizeof(e));
+	printf("Size of a char: %ld byte(s)\n", sizeof(a));
+	printf("Size of an int: %li byte(s)\n", sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %li bytes(s)\n", sizeof(d));
+	printf("Size of a float: %ld bytes(s)\n", sizeof(e));
 	}
 	return (0);
 }
